Split sender into lib and bin
