+++ utils/sender/Cargo.toml
@@ -17,8 +17,13 @@ tokio = { version = "1.28.2", features = ["rt-multi-thread", "macros", "time"] }
 codec = { package = "parity-scale-codec", version = "3.5.0", default-features = false, features = ["derive", "full", "bit-vec"] }
 futures = "0.3.28"
 
+[lib]
+name = "sender_lib"
+path = "src/lib.rs"
+
 [[bin]]
 name = "sender"
+path = "src/main.rs"
 
 [features]
 tick = ["utils/tick"]
